pyproject.toml
[project]
name = "att-lstm"
version = "0.1.0"
description = "EPS Torque Prediction using LSTM with Attention Mechanisms"
readme = "docs/README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "torch>=2.0.0",
    "pytorch-lightning>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "scienceplots>=2.1.0",
    "optuna>=3.0.0",
    "tqdm>=4.65.0",
    "tensorboard>=2.14.0",
    "pyyaml>=6.0",
    "thop>=0.1.1",
]

[tool.uv.sources]
torch = { index = "pytorch-cu124" }

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
